requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.0
python-dotenv==1.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
